+++ ai/train/isaac_runner.py
@@ -47,7 +47,7 @@ simulation_app = app_launcher.app
 # ── Imports that require Isaac Sim to be running ──────────────────────
 
 import gymnasium as gym  # noqa: E402
-import numpy as np  # noqa: E402
+
 import torch  # noqa: E402
 
 import isaaclab_tasks  # noqa: E402, F401
@@ -628,361 +628,9 @@ def _eval_rsl_rl(cfg, isaac_task, checkpoint_path, num_envs, num_episodes):
     env.close()
 
 
-# ── Trajectory Export ─────────────────────────────────────────────────
+# ── Trajectory Export (see trajectory_export/ package) ────────────────
 
-def _detect_gait_cycle(positions: np.ndarray) -> int | None:
-    """Find gait period via autocorrelation on the first joint's position."""
-    sig = positions[:, 0].copy()
-    sig -= sig.mean()
-    std = sig.std()
-    if std < 1e-8:
-        return None
-    sig /= std
-
-    n = len(sig)
-    autocorr = np.correlate(sig, sig, mode="full")[n - 1:]
-    autocorr /= autocorr[0]
-
-    min_lag = max(5, n // 50)
-    for i in range(1, len(autocorr)):
-        if autocorr[i] < 0:
-            min_lag = max(min_lag, i)
-            break
-
-    search = autocorr[min_lag:]
-    if len(search) == 0:
-        return None
-    peak_idx = int(np.argmax(search)) + min_lag
-    if autocorr[peak_idx] < 0.3:
-        return None
-    return peak_idx
-
-
-def _write_trajectory_pbtxt(
-    path: str,
-    topics: list[str],
-    timestamps: np.ndarray,
-    data: np.ndarray,
-    action_type: str,
-    node_id: int,
-) -> None:
-    """Write a .pbtxt trajectories block matching TrajectoryPublisher format."""
-    lines = [
-        "trajectories {",
-        "  single_trajectories {",
-        "    node {",
-        f"      id: {node_id}",
-        "      node_type: TRAJECTORY_PUBLISHER",
-        "      qos_setting { depth: 10 }",
-    ]
-    for topic in topics:
-        lines.append(
-            f'      publishers {{ ros2_data_type: FLOAT32  topic: "{topic}" }}'
-        )
-    lines.append("    }")
-    lines.append("    trajectory {")
-    num_steps = data.shape[0]
-    for t in range(num_steps):
-        for j, topic in enumerate(topics):
-            lines.append(
-                f"      waypoints {{ timestamp_sec: {timestamps[t]:.6f}  "
-                f'topic: "{topic}"  '
-                f"action {{ {action_type}: {data[t, j]:.6f} }} }}"
-            )
-    lines.append("    }")
-    lines.append("  }")
-    lines.append("}")
-    with open(path, "w") as f:
-        f.write("\n".join(lines) + "\n")
-    print(f"[Joshua/Isaac] Wrote {path}")
-
-
-def _export_trajectory_data(
-    positions: np.ndarray,
-    actions: np.ndarray,
-    step_dt: float,
-    cfg: dict,
-    joint_names: list[str],
-) -> None:
-    """Process recorded data, detect cycle, and write output files."""
-    output_dir = cfg.get("output_dir", "")
-    if not output_dir:
-        output_dir = os.path.join(
-            cfg.get("checkpoint_dir", "/tmp/joshua_checkpoints"),
-            "trajectory_export",
-        )
-    os.makedirs(output_dir, exist_ok=True)
-
-    detect_cycle = cfg.get("detect_cycle", False)
-    node_id = cfg.get("trajectory_node_id", 1)
-    mappings = cfg.get("joint_topic_mappings", [])
-
-    topic_map = {m["joint_name"]: m["topic"] for m in mappings}
-
-    export_indices = []
-    export_topics = []
-    for i, name in enumerate(joint_names):
-        if name in topic_map:
-            export_indices.append(i)
-            export_topics.append(topic_map[name])
-        else:
-            print(f"[Joshua/Isaac] WARNING: no topic mapping for joint "
-                  f"'{name}', skipping")
-
-    if not export_indices:
-        raise ValueError(
-            f"No joints matched joint_topic_mappings. "
-            f"Available joints: {joint_names}"
-        )
-
-    pos_data = positions[:, export_indices]
-    act_data = actions[:, export_indices]
-
-    if detect_cycle:
-        period = _detect_gait_cycle(pos_data)
-        if period and period < len(pos_data):
-            print(f"[Joshua/Isaac] Detected gait cycle: {period} steps "
-                  f"({period * step_dt:.3f}s)")
-            pos_data = pos_data[:period]
-            act_data = act_data[:period]
-        else:
-            print("[Joshua/Isaac] Could not detect cycle, using full recording")
-
-    num_steps = len(pos_data)
-    timestamps = np.arange(num_steps) * step_dt
-
-    print(f"[Joshua/Isaac] Exporting {num_steps} steps, "
-          f"{len(export_topics)} joints, "
-          f"duration={timestamps[-1]:.3f}s")
-
-    _write_trajectory_pbtxt(
-        os.path.join(output_dir, "trajectory_position.pbtxt"),
-        export_topics, timestamps, pos_data, "position", node_id,
-    )
-    _write_trajectory_pbtxt(
-        os.path.join(output_dir, "trajectory_torque.pbtxt"),
-        export_topics, timestamps, act_data, "torque", node_id,
-    )
-
-    npy_pos_path = os.path.join(output_dir, "trajectory_position.npy")
-    np.save(npy_pos_path, pos_data)
-    print(f"[Joshua/Isaac] Wrote {npy_pos_path} {pos_data.shape}")
-
-    npy_act_path = os.path.join(output_dir, "trajectory_torque.npy")
-    np.save(npy_act_path, act_data)
-    print(f"[Joshua/Isaac] Wrote {npy_act_path} {act_data.shape}")
-
-    meta = {
-        "joint_names": [joint_names[i] for i in export_indices],
-        "topics": export_topics,
-        "num_steps": num_steps,
-        "step_dt_s": step_dt,
-        "total_duration_s": float(timestamps[-1]),
-        "position_units": "observed joint positions (radians)",
-        "torque_units": "raw policy action outputs (pre-action-scale)",
-    }
-    meta_path = os.path.join(output_dir, "trajectory_meta.json")
-    with open(meta_path, "w") as f:
-        json.dump(meta, f, indent=2)
-    print(f"[Joshua/Isaac] Wrote {meta_path}")
-
-
-def _trajectory_export_rsl_rl(cfg: dict) -> None:
-    """Export trajectory from an RSL-RL checkpoint."""
-    from rsl_rl.runners import OnPolicyRunner
-    from isaaclab_rl.rsl_rl import RslRlVecEnvWrapper
-
-    isaac_task = _resolve_task(cfg)
-    checkpoint_path = cfg["checkpoint_path"]
-    device = "cuda" if torch.cuda.is_available() else "cpu"
-
-    env_cfg = parse_env_cfg(isaac_task, num_envs=1)
-    _apply_env_overrides(env_cfg, cfg)
-    env = gym.make(isaac_task, cfg=env_cfg)
-
-    raw_env = env.unwrapped
-    joint_names = list(raw_env.scene["robot"].data.joint_names)
-    step_dt = raw_env.cfg.sim.dt * raw_env.cfg.decimation
-
-    env = RslRlVecEnvWrapper(env)
-
-    agent_cfg = load_cfg_from_registry(isaac_task, "rsl_rl_cfg_entry_point")
-    _apply_agent_overrides(agent_cfg, cfg)
-    agent_cfg.device = device
-
-    runner = OnPolicyRunner(
-        env, agent_cfg.to_dict(), log_dir=None, device=device,
-    )
-    runner.load(checkpoint_path)
-
-    warmup_steps = cfg.get("warmup_steps", 200)
-    num_record_steps = cfg.get("num_record_steps", 1000)
-    num_joints = len(joint_names)
-
-    print(f"[Joshua/Isaac] Joints ({num_joints}): {joint_names}")
-    print(f"[Joshua/Isaac] dt={step_dt:.6f}s  "
-          f"warmup={warmup_steps}  record={num_record_steps}")
-
-    obs = env.get_observations()
-    for _ in range(warmup_steps):
-        with torch.inference_mode():
-            actions = runner.alg.act(obs)
-        obs, _, _, dones, _ = env.step(actions)
-
-    all_positions = np.zeros((num_record_steps, num_joints))
-    all_actions = np.zeros((num_record_steps, num_joints))
-
-    for t in range(num_record_steps):
-        with torch.inference_mode():
-            actions = runner.alg.act(obs)
-        obs, _, _, dones, _ = env.step(actions)
-        all_positions[t] = raw_env.scene["robot"].data.joint_pos[0].cpu().numpy()
-        all_actions[t] = actions[0].cpu().numpy()
-
-    env.close()
-    _export_trajectory_data(
-        all_positions, all_actions, step_dt, cfg, joint_names)
-
-
-def _trajectory_export_skrl(cfg: dict) -> None:
-    """Export trajectory from a skrl checkpoint."""
-    import skrl  # noqa: F401
-    from skrl.agents.torch.ppo import PPO, PPO_DEFAULT_CONFIG
-    from skrl.envs.wrappers.torch import wrap_env
-    from skrl.memories.torch import RandomMemory
-    from skrl.models.torch import DeterministicMixin, GaussianMixin, Model
-
-    isaac_task = _resolve_task(cfg)
-    checkpoint_path = cfg["checkpoint_path"]
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-    env_cfg = parse_env_cfg(isaac_task, num_envs=1)
-    _apply_env_overrides(env_cfg, cfg)
-    env = gym.make(isaac_task, cfg=env_cfg)
-
-    raw_env = env.unwrapped
-    joint_names = list(raw_env.scene["robot"].data.joint_names)
-    step_dt = raw_env.cfg.sim.dt * raw_env.cfg.decimation
-
-    env = wrap_env(env, wrapper="isaaclab")
-
-    try:
-        agent_cfg = load_cfg_from_registry(isaac_task, "rsl_rl_cfg_entry_point")
-    except KeyError:
-        agent_cfg = None
-
-    if agent_cfg is not None:
-        _apply_agent_overrides(agent_cfg, cfg)
-
-    obs_size = env.observation_space.shape[-1]
-    action_size = env.action_space.shape[-1]
-    hidden = agent_cfg.policy.actor_hidden_dims if agent_cfg else [256, 256]
-
-    class Policy(GaussianMixin, Model):
-        def __init__(self, observation_space, action_space, dev, **kwargs):
-            Model.__init__(self, observation_space, action_space, dev)
-            GaussianMixin.__init__(self, min_log_std=-20.0, max_log_std=2.0)
-            layers = []
-            in_dim = obs_size
-            for h in hidden:
-                layers += [torch.nn.Linear(in_dim, h), torch.nn.ELU()]
-                in_dim = h
-            layers.append(torch.nn.Linear(in_dim, action_size))
-            self.net = torch.nn.Sequential(*layers)
-            self.log_std_parameter = torch.nn.Parameter(
-                torch.zeros(action_size))
-
-        def compute(self, inputs, role=""):
-            return self.net(inputs["states"]), self.log_std_parameter, {}
-
-    class Value(DeterministicMixin, Model):
-        def __init__(self, observation_space, action_space, dev, **kwargs):
-            Model.__init__(self, observation_space, action_space, dev)
-            DeterministicMixin.__init__(self)
-            c_hidden = (agent_cfg.policy.critic_hidden_dims
-                        if agent_cfg else [256, 256])
-            layers = []
-            in_dim = obs_size
-            for h in c_hidden:
-                layers += [torch.nn.Linear(in_dim, h), torch.nn.ELU()]
-                in_dim = h
-            layers.append(torch.nn.Linear(in_dim, 1))
-            self.net = torch.nn.Sequential(*layers)
-
-        def compute(self, inputs, role=""):
-            return self.net(inputs["states"]), {}
-
-    models = {
-        "policy": Policy(env.observation_space, env.action_space, device),
-        "value": Value(env.observation_space, env.action_space, device),
-    }
-
-    rollout_steps = agent_cfg.num_steps_per_env if agent_cfg else 24
-    memory = RandomMemory(
-        memory_size=rollout_steps, num_envs=1, device=device)
-
-    ppo_cfg = PPO_DEFAULT_CONFIG.copy()
-    agent = PPO(
-        models=models,
-        memory=memory,
-        cfg=ppo_cfg,
-        observation_space=env.observation_space,
-        action_space=env.action_space,
-        device=device,
-    )
-    agent.load(checkpoint_path)
-    agent.set_running_mode("eval")
-
-    warmup_steps = cfg.get("warmup_steps", 200)
-    num_record_steps = cfg.get("num_record_steps", 1000)
-    num_joints = len(joint_names)
-
-    print(f"[Joshua/Isaac] Joints ({num_joints}): {joint_names}")
-    print(f"[Joshua/Isaac] dt={step_dt:.6f}s  "
-          f"warmup={warmup_steps}  record={num_record_steps}")
-
-    obs, _ = env.reset()
-    for _ in range(warmup_steps):
-        with torch.inference_mode():
-            outputs = agent.act(obs, timestep=0, timesteps=0)
-            actions = outputs[-1].get("mean_actions", outputs[0])
-            obs, _, _, _, _ = env.step(actions)
-
-    all_positions = np.zeros((num_record_steps, num_joints))
-    all_actions = np.zeros((num_record_steps, num_joints))
-
-    for t in range(num_record_steps):
-        with torch.inference_mode():
-            outputs = agent.act(obs, timestep=0, timesteps=0)
-            actions = outputs[-1].get("mean_actions", outputs[0])
-            obs, _, _, _, _ = env.step(actions)
-        all_positions[t] = raw_env.scene["robot"].data.joint_pos[0].cpu().numpy()
-        all_actions[t] = actions[0].cpu().numpy()
-
-    env.close()
-    _export_trajectory_data(
-        all_positions, all_actions, step_dt, cfg, joint_names)
-
-
-def _run_trajectory_export(cfg: dict) -> None:
-    """Export a trained policy as a constant trajectory."""
-    algorithm = cfg.get("algorithm", "rsl_rl")
-    checkpoint_path = cfg.get("checkpoint_path", "")
-    if not checkpoint_path:
-        raise ValueError("checkpoint_path is required for trajectory export")
-    if not os.path.isfile(checkpoint_path):
-        raise FileNotFoundError(f"Checkpoint not found: {checkpoint_path}")
-
-    print(f"[Joshua/Isaac] Trajectory export ({algorithm})")
-    print(f"[Joshua/Isaac] Checkpoint: {checkpoint_path}")
-
-    if algorithm == "rsl_rl":
-        _trajectory_export_rsl_rl(cfg)
-    elif algorithm == "skrl":
-        _trajectory_export_skrl(cfg)
-    else:
-        raise ValueError(
-            f"Unknown algorithm '{algorithm}' for trajectory export")
+from trajectory_export import run_trajectory_export as _run_trajectory_export  # noqa: E402
 
 
 # ── Entry point ──────────────────────────────────────────────────────
